found the problem in pressing delete
