+++ components/editTask.js
@@ -1,70 +1,6 @@
 import { dateConversion } from "./submitTasks.js";
 import { createForm } from "./tasks.js";
 
-export function editTitle() {
-  const getProjectTitle = document.querySelectorAll(".project h3");
-  getProjectTitle.forEach(projectTitle => {
-    projectTitle.style.cursor = "pointer";
-    
-    projectTitle.addEventListener("click", (e) => {
-      var counter = e.currentTarget.parentElement.id.split('-')[1];
-      deleteTitle(counter);
-      projectTitle.contentEditable = "true";
-      projectTitle.focus();
-    });
-
-    projectTitle.addEventListener("blur", (e) => {projectTitle.contentEditable = "false";
-      var counter = e.currentTarget.parentElement.id.split('-')[1];
-      viewAddButton(counter);
-      removeDeleteButton(counter);
-    });
-
-    projectTitle.addEventListener("keydown", (e) => {
-      if (e.key === 'Enter') {
-        projectTitle.contentEditable = "false";
-      }
-    });
-  });
-}
-
-function deleteTitle(counter) {
-    removeAddButton(counter);
-    createDeleteButton(counter);
-}
-
-function createDeleteButton(counter) {
-  const addButtons = document.querySelectorAll('.add-task-button');
-  addButtons.forEach(addButton => {
-    let deleteButton = addButton.nextElementSibling;
-    if (!deleteButton || deleteButton.name !== "deleteButton") {
-      deleteButton = document.createElement("button");
-      deleteButton.innerText = 'Delete Project';
-      deleteButton.id = `delete-button-${counter}`;
-      deleteButton.name = "deleteButton";
-      deleteButton.className = 'delete-button';
-      addButton.after(deleteButton);
-    }
-    else if (deleteButton.style.display === 'none') {
-      deleteButton.style.display = 'inherit';
-    }
-  });
-}
-
-function removeAddButton(counter) {
-  const addButton = document.getElementById(`addTaskButton-${counter}`);
-  addButton.style.display = 'none';
-}
-
-function removeDeleteButton(counter) {
-  const removeButton = document.getElementById(`delete-button-${counter}`);
-  removeButton.style.display = 'none';
-}
-
-function viewAddButton(counter) {
-  const addButton = document.getElementById(`addTaskButton-${counter}`);
-  addButton.style.display = 'inherit';
-}
-
 export default function editTaskHandler(buttonCounter) {
     editSelection(buttonCounter);
 }
